Use Numpy to improve performance
Rather than three nested loops, we now have one loop over a two-dimensional matrix. The arithmetic game runs enormously quicker than the non-numpy version -- a fraction of a second instead of minutes. The geometric game is much quicker, but still takes several seconds (rather than a substantial fraction of an hour before). What seems to slow the geometric game down is the minimum function, which executes every draw, and which allocates a large matrix on the heap each time.